init-system-helpers: drop update-rc.d
As update-rc.d command provided by update-rc.d package is used by
update-rc.d class. The class and package are deeply connected. So, it is
not a good idea to override the command.

This also drop dh_systemd sub-package, as it only contains empty dirs,
and drops ALTERNATIVE_... environment variables as e.g. service command
is not provided by any other packages.

This moves service command and friends from /usr/bin to /usr/sbin, as
the commands should always invoked by administrators.

Moreover, this drops the header as now the recipe is far from the original.