libpwquality: Add libpwquality license file
According to the debian copyright file[1], the libpwquality has dual
license which the one is libpwquality and the other is GPL-2+.

The libpwquality license is not common license so that it doesn't exist
in meta/files/common-licenses in poky.

Therefore, get the license from debian's copyright file[1], then added
it to custom-licenses directory.

1. https://metadata.ftp-master.debian.org/changelogs//main/libp/libpwquality/libpwquality_1.4.0-3_copyright

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>