glibc: glibc-locale: Add workaround to prevent random host-user-contaminated warning
Added workaround to prevent host-user-contaminated warning.
This issue was discussed in the openembedded-core mailing list[1,2].
On the maling list, developers are talking a workaround which is set user and
group in do_prep_locale_tree() task. This workaround is not merged into
any branches yet. So, we will apply this workaround for us.

1.https://lore.kernel.org/openembedded-core/20220616055414.42319-1-muhammad_hamza@mentor.com/T/#t
2.https://patches.linaro.org/project/oe-core/patch/20190207003537.7135-1-raj.khem@gmail.com/

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>