apparmor: Don't remove sysvinit scripts
apparmor.service is a wrapper for /etc/init.d/apparmor.
However, bitbake removes sysvinit-scripts at build time in a systemd environment.
(On the other hand, in a sysvinit environment, bitbake removes the systemd.service file.)
This caused a issue that apparmor could not be started when systemd is enabled.

To solve this issue, this patch will not make bitbake delete sysvinit-scripts in apparmor build.

Fix:
| Nov 16 04:25:25 qemuarm64 systemd[1]: Starting AppArmor initialization...
| Nov 16 04:25:25 qemuarm64 systemd[171]: apparmor.service: Failed to execute command: No such file or directory
| Nov 16 04:25:25 qemuarm64 systemd[171]: apparmor.service: Failed at step EXEC spawning /etc/init.d/apparmor: No such file or directory
| Nov 16 04:25:25 qemuarm64 systemd[1]: apparmor.service: Main process exited, code=exited, status=203/EXEC
| Nov 16 04:25:25 qemuarm64 systemd[1]: apparmor.service: Failed with result 'exit-code'.
| Nov 16 04:25:25 qemuarm64 systemd[1]: Failed to start AppArmor initialization.