clamav: workaround: Update from 0.103.9+dfsg-0+deb10u1 to 0.103.12+dfsg-0+deb9u1
This version fixes following CVEs.

- CVE-2024-20505
- CVE-2024-20506

ELA: ELA-1268-1

The ClamAV package for Debian 10 has been upgraded and the major
version has changed. To build the new ClamAV package needs to prepare
rust-toolchains and a newer cmake.

This commit enables to install the ClamAV package for Debian 9 which
has been already fixed the above CVEs, as a workaround.

With the version update, the package directory name in tarball has
been changed to not include the repack string such as "+dfsg".
Fixed DEBIAN_UNPACK_DIR to match this directory name change.

Signed-off-by: Hirotaka Motai <hirotaka.motai@miraclelinux.com>